echuu-web/backend/requirements.txt
# ECHUU Web Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
